fix MessageAttributesTest
