+++ amazon/sns/client/src/main/kotlin/org/http4k/connect/amazon/sns/model/SNSMessageAttribute.kt
@@ -2,6 +2,7 @@ package org.http4k.connect.amazon.sns.model
 
 import org.http4k.connect.amazon.core.model.DataType
 import org.http4k.connect.amazon.core.model.MessageFields
+import org.http4k.connect.amazon.core.model.MessageFieldsDto
 
 sealed class SNSMessageAttribute(
     override val name: String,
@@ -25,6 +26,12 @@ sealed class SNSMessageAttribute(
     ) : SNSMessageAttribute(name, category, dataType) {
         override fun toCustomFields(index: Int) =
             mapOf("$category.entry.$index.Value.${typePrefix}Value" to value)
+
+        override fun toDto() = MessageFieldsDto(
+            dataType = dataType,
+            stringValue = if (dataType != DataType.Binary) value else null,
+            binaryValue = if (dataType == DataType.Binary) value else null
+        )
     }
 
     internal class ListValue(
@@ -40,5 +47,11 @@ sealed class SNSMessageAttribute(
             values.mapIndexed { i, it ->
                 "$category.entry.$index.Value.${typePrefix}Value.${i + 1}" to it
             }.toMap()
+
+        override fun toDto() = MessageFieldsDto(
+            dataType = dataType,
+            stringListValues = if (dataType != DataType.Binary) values else null,
+            binaryListValues = if (dataType == DataType.Binary) values else null
+        )
     }
 }
