+++ amazon/sqs/client/src/main/kotlin/org/http4k/connect/amazon/sqs/action/SendMessageBatch.kt
@@ -6,7 +6,6 @@ import org.http4k.connect.amazon.sqs.model.BatchResultErrorEntry
 import org.http4k.connect.amazon.sqs.model.MessageAttribute
 import org.http4k.connect.amazon.sqs.model.MessageSystemAttribute
 import org.http4k.connect.amazon.sqs.model.SQSMessageId
-import org.http4k.connect.amazon.sqs.model.toDto
 import org.http4k.core.Uri
 import se.ansman.kotshi.JsonSerializable
 import com.squareup.moshi.Json
