+++ amazon/sqs/client/src/main/kotlin/org/http4k/connect/amazon/sqs/model/SQSMessage.kt
@@ -2,7 +2,9 @@ package org.http4k.connect.amazon.sqs.model
 
 import se.ansman.kotshi.JsonSerializable
 import com.squareup.moshi.Json
+import org.http4k.connect.amazon.core.model.DataType
 import org.http4k.connect.amazon.core.model.MessageFieldsDto
+import org.http4k.connect.model.Base64Blob
 
 @JsonSerializable
 data class SQSMessage(
@@ -27,4 +29,17 @@ data class SQSMessage(
     )
 
     val attributes get() = messageAttributes.map { (name, value) -> value.toSqs(name) }
+}
+
+internal fun MessageFieldsDto.toSqs(name: String) = when(dataType) {
+    DataType.String, DataType.Number -> if (stringListValues != null) {
+        MessageAttribute(name, stringListValues.orEmpty(), dataType)
+    } else {
+        MessageAttribute(name, stringValue!!, dataType)
+    }
+    DataType.Binary -> if (binaryListValues != null) {
+        MessageAttribute(name, binaryListValues.orEmpty().map(Base64Blob::of))
+    } else {
+        MessageAttribute(name, Base64Blob.of(binaryValue!!))
+    }
 }
